(PC-33397)[BO] chore: move pages meant to testing only in a dev subfolder
