+++ api/src/pcapi/routes/backoffice/dev/forms.py
@@ -1,3 +1,4 @@
+from flask_wtf import FlaskForm
 import wtforms
 
 from pcapi.core.fraud.models import UBBLE_OK_REASON_CODE
@@ -7,6 +8,30 @@ from pcapi.core.users.generator import GeneratedIdProvider
 from pcapi.core.users.generator import GeneratedSubscriptionStep
 from pcapi.routes.backoffice.forms import fields
 from pcapi.routes.backoffice.forms import utils
+
+from ..forms import fields
+
+
+class SimpleComponentsForm(FlaskForm):
+    checkbox = fields.PCCheckboxField("checkbox")
+    stringfield = fields.PCStringField("string field")
+    password = fields.PCPasswordField("password")
+    postal_code = fields.PCPostalCodeField("code postal")
+    text_area = fields.PCTextareaField("text area")
+    interger = fields.PCIntegerField("nombre entier")
+    decimal = fields.PCDecimalField("nombre décimal")
+    date_time = fields.PCDateTimeField("date et heure")
+    date = fields.PCDateField("date")
+    date_range = fields.PCDateRangeField("intervalle de dates")
+    select = fields.PCSelectWithPlaceholderValueField("un seul choix", choices=((1, "choix 1"), (2, "choix 2")))
+    multiselect = fields.PCSelectMultipleField("plusieurs choix", choices=((1, "choix 1"), (2, "choix 2")))
+    tomselect = fields.PCTomSelectField(
+        "autocomplete",
+        choices=[],
+        validate_choice=False,
+        endpoint="backoffice_web.autocomplete_venues",
+    )
+    switch = fields.PCSwitchBooleanField("interrupteur")
 
 
 class UserGeneratorForm(utils.PCForm):
