+++ api/src/pcapi/routes/backoffice/templates/dev/users_deletion.html
@@ -3,8 +3,8 @@
   <div class="pt-3 px-5">
     <h2 class="fw-light">Suppression d'utilisateur de test</h2>
     <div class="col">
-      <form action="{{ url_for('backoffice_web.delete_user') }}"
-            name="{{ url_for('backoffice_web.delete_user') | action_to_name }}"
+      <form action="{{ url_for('backoffice_web.dev.delete_user') }}"
+            name="{{ url_for('backoffice_web.dev.delete_user') | action_to_name }}"
             method="post"
             data-turbo="false">
         {{ form.csrf_token }}
