+++ api/src/pcapi/routes/backoffice/templates/dev/users_generator.html
@@ -52,7 +52,7 @@
                   </div>
                 </a>
               {% endif %}
-              {{ build_modal_form("configure-ubble", url_for("backoffice_web.configure_ubble_v2_response", user_id=user.id) , ubble_configuration_form, "Configurer les réponses d'Ubble v2", "Configurer les réponses d'Ubble v2", "Enregistrer") }}
+              {{ build_modal_form("configure-ubble", url_for("backoffice_web.dev.configure_ubble_v2_response", user_id=user.id) , ubble_configuration_form, "Configurer les réponses d'Ubble v2", "Configurer les réponses d'Ubble v2", "Enregistrer") }}
             </div>
             <hr>
             <div class="d-flex justify-content-start">
