+++ api/src/pcapi/settings.py
@@ -536,6 +536,7 @@ BACKOFFICE_LOCAL_USER_EMAIL = os.environ.get("BACKOFFICE_LOCAL_USER_EMAIL", "adm
 BACKOFFICE_LOGIN_WITHOUT_CREDENTIALS = bool(int(os.environ.get("BACKOFFICE_LOGIN_WITHOUT_CREDENTIALS", 0)))
 BACKOFFICE_ROLES_WITHOUT_GOOGLE_GROUPS = bool(int(os.environ.get("BACKOFFICE_ROLES_WITHOUT_GOOGLE_GROUPS", 0)))
 BACKOFFICE_WATCH_SCSS_CHANGE = bool(int(os.environ.get("BACKOFFICE_WATCH_SCSS_CHANGE", 0)))
+ENABLE_BO_COMPONENT_PAGE = bool(int(os.environ.get("ENABLE_BO_COMPONENT_PAGE", 0)))
 
 # SIRENE
 SIRENE_BACKEND = os.environ.get("SIRENE_BACKEND")
